// src/types/inventories.ts
export interface FacetItem {
  count: number;
  name: string;
  order?: number;
  slug: string;
}
// ...
export type item_type =
  | "armor"
  | "weapon"
  | "blueprint"
  | "medical"
  | "trade-good"
  | "mission"
  | "ration"
  | "food"
  | "event";

export type rarity = "common" | "uncommon" | "rare" | "epic";

export interface InventoryFilter {
  item_type?: item_type[];
  rarity?: rarity[];
  tier?: number[];
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  [key: string]: any;
}